Move functionality into game.js file and add length when snake eats apple
