+++ lib/Game.js
@@ -1,0 +1,98 @@
+var Head = require('./Head.js');
+var Fruit = require('./Fruit.js');
+
+var head = new Head(60, 40, 'rgb(0, 250, 0)', 0, 0);
+var fruit = new Fruit();
+
+var canvas = document.querySelector('#game');
+var context = canvas.getContext('2d');
+var isGameOver = false;
+var blocks = [head];
+var fruits = [fruit];
+var score = 0;
+var scoreValue = document.querySelector('.score');
+
+
+requestAnimationFrame(gameLoop);
+
+window.addEventListener('keydown', changeDirection);
+
+
+function gameLoop () {
+  context.clearRect(0, 0, canvas.width, canvas.height); 
+
+  if (isGameOver) {
+    return alert('game over');
+
+  } else {
+    animate();
+  }
+
+  requestAnimationFrame(gameLoop);
+}
+
+function isCollidingWith(object) {
+  return !(
+    blocks[0].x + blocks[0].width < object.x ||
+    blocks[0].y + blocks[0].height < object.y ||
+    blocks[0].x > object.x + object.width || 
+    blocks[0].y > object.y + object.height
+  );
+}
+
+function animate() {
+  
+  fruits.forEach(function(fruit){
+    fruit.draw(context);
+  })
+  
+  blocks.forEach(function (block) {
+    block.draw(context);
+    block.move();
+    
+  
+    blocks.forEach(function (block) {
+      if (block.x > 797 || block.x === 0 || block.y > 647 || block.y === -2.5) {
+        isGameOver = true;
+      } else if (isCollidingWith(fruit)) {
+        fruit.x = Math.floor(Math.random() * (800 - 0) / 20) * 20 + 0;
+        fruit.y = Math.floor(Math.random() * (600 - 0) / 20) * 20 + 0;
+        score++;
+        scoreValue.innerText = score;
+        growSnake(block);
+      }
+    })
+  });
+}
+
+function growSnake(block) {
+  if (head.dy === 1) {
+    blocks.push(new Head(block.x, block.y - 20, 'rgb(0, 250, 0)', 0, 0));
+    } else if (head.dy === -1) {
+      blocks.push(new Head(block.x, block.y + 20, 'rgb(0, 250, 0)', 0, 0));
+    } else if (head.dx === 1) {
+      blocks.push(new Head(block.x - 20, block.y, 'rgb(0, 250, 0)', 0, 0));
+    } else if (head.dx === -1) {
+      blocks.push(new Head(block.x + 20, block.y, 'rgb(0, 250, 0)', 0, 0));
+    }
+}
+
+function changeDirection(e) {
+  if (e.keyCode === 38 && head.dy !== 1) {
+    head.x -= head.x % 20;
+    head.dy = -1;
+    head.dx = 0;
+  } else if (e.keyCode === 39 && head.dx !== -1) {
+    head.y -= head.y % 20;
+    head.dx = 1;
+    head.dy = 0
+  } else if (e.keyCode === 40 && head.dy !== -1) {
+    head.x -= head.x % 20;
+    head.dy = 1;
+    head.dx = 0;
+  } else if (e.keyCode === 37 && head.dx !== 1) {
+    head.y -= head.y % 20;
+    head.dx = -1;
+    head.dy = 0;
+  }
+}
