+++ lib/GamePiece.js
@@ -1,5 +1,5 @@
 module.exports = class GamePiece {
-  constructor(x, y, color, height, width) {
+  constructor(x, y, color) {
     this.x = x;
     this.y = y;
     this.color = color;
@@ -7,19 +7,10 @@ module.exports = class GamePiece {
     this.width = 19.9;
   }
 
-  isCollidingWith(object) {
-    return !(
-      this.x + this.width < object.x ||
-      this.y + this.height < object.y ||
-      this.x > object.x + object.width || 
-      this.y > object.y + object.height
-    );
-  }
-
   draw(ctx) {
     const { x, y, height, width, color } = this;
 
     ctx.fillStyle = color;
-    ctx.fillRect(x, y, height, width);
+    ctx.fillRect(x, y, width, height);
   }
 }
