+++ lib/Head.js
@@ -1,12 +1,10 @@
 var GamePiece = require('./GamePiece.js');
 
-// window.addEventListener('keydown', changeDirection);
-
 module.exports = class Head extends GamePiece {
   constructor(x, y, color, dx, dy) {
     super(x, y, color);
-    this.dx = dx;   // direction: 1 === right, -1 === left
-    this.dxv = 6; // velocity / speed
+    this.dx = dx;
+    this.dxv = 6;
     this.dy = dy;
     this.dyv = 6;
   }
@@ -18,19 +16,3 @@ module.exports = class Head extends GamePiece {
     }
   };
 }
-  // function changeDirection(e) {
-  //   if (e.keyCode === 38) {
-  //     head.dy = -1;
-  //     head.dx = 0;
-  //     console.log(head.dy);
-  //   } else if (e.keyCode === 39) {
-  //     head.dx = 1;
-  //     head.dy = 0;
-  //   } else if (e.keyCode === 40) {
-  //     head.dy = 1;
-  //     head.dx = 0;
-  //   } else if (e.keyCode === 37) {
-  //     head.dx = -1;
-  //     head.dy = 0;
-  //   }
-  // }
